+++ src/index.ts
@@ -32,6 +32,12 @@ export type {
 export type { CombinedFilter } from './query-builder/types/filter/combined-filter.type';
 export type { Guid } from './query-builder/types/utils/util.types';
 export type { ExpandFields } from './query-builder/types/expand/expand-fields.type';
+export type {
+    ExpandInput,
+    ExpandSubQueryOptions,
+    ExpandWithSubQuery,
+    TopLevelExpandFields,
+} from './query-builder/types/expand/expand-descriptor.type';
 
 export { isCombinedFilter } from './query-builder/utils/filter/combined-filter-util';
 export { isQueryFilter } from './query-builder/utils/filter/is-query-filter-util';